+++ ui/src/App.vue
@@ -1,11 +1,12 @@
 <template>
-  <router-view />
+  <MainLayout />
+  
 </template>
 
 <script setup lang="ts">
+import MainLayout from './layouts/MainLayout.vue'
 </script>
 
 <style>
 html, body, #app { height: 100%; margin: 0; }
 </style>
-